refactor: convert to using stdint

+++ include/opponent.h
@@ -1,12 +1,13 @@
 #pragma once
 
 #include <stdbool.h>
+#include <stdint.h>
 
 #include "events.h"
 
 typedef struct Opponent {
     bool finished;
-    int milk_consumed;
+    uint8_t milk_consumed;
 } Opponent;
 
 Opponent* opponent_create(void);
